poc/access-odbc-python/pyproject.toml
[tool.poetry]
name = "access-odbc-python"
version = "0.1.0"
description = ""
authors = ["Wickes1 <27881570+wickes1@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "access_odbc_python"}]

[tool.poetry.dependencies]
python = "^3.11"
pyodbc = "^5.1.0"
pandas = "^2.2.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
